requirements.txt
pytest>=7.0.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0
ruff>=0.1.0
psycopg2-binary>=2.9.0
confluent-kafka>=2.0.0
python-dotenv>=1.0.0
git+https://github.com/J-Quants/jquants-api-client-python.git